pull out VerticalDivider
