+++ context/app/static/js/components/Detail/SummaryItem.jsx
@@ -1,15 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
 import Typography from '@material-ui/core/Typography';
-import Divider from '@material-ui/core/Divider';
-
-const VerticalDivider = styled(Divider)`
-  margin-left: ${(props) => props.theme.spacing(1)}px;
-  margin-right: ${(props) => props.theme.spacing(1)}px;
-  height: 15px;
-  background-color: ${(props) => props.theme.palette.text.primary};
-  align-self: center;
-`;
 
 const StyledDiv = styled.div`
   display: flex;
@@ -23,7 +14,6 @@ function SummaryItem(props) {
       <Typography variant="h6" component="p">
         {children}
       </Typography>
-      <VerticalDivider orientation="vertical" flexItem />
     </StyledDiv>
   );
 }
